move navigation to its own folder
