+++ frontend/packages/ux-editor/src/components/Settings/SettingsNavigation/SettingsNavigation.test.tsx
