+++ frontend/packages/ux-editor/src/components/Settings/SettingsNavigation/SettingsNavigation.tsx
@@ -4,10 +4,10 @@ import type { TaskNavigationGroup } from 'app-shared/types/api/dto/TaskNavigatio
 import { StudioAlert, StudioSpinner } from '@studio/components-legacy';
 import React, { type ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
-import { isTaskReceipt, getTaskIcon, taskNavigationType } from './SettingsUtils';
+import { isTaskReceipt, getTaskIcon, taskNavigationType } from '../SettingsUtils';
 import { PadlockLockedFillIcon } from '@studio/icons';
 import classes from './SettingsNavigation.module.css';
-import { TasksTable } from './SettingsNavigation/TasksTable';
+import { TasksTable } from '../SettingsNavigation/TasksTable';
 
 export const SettingsNavigation = (): ReactElement => {
   const { t } = useTranslation();
